Move Pipeline code to own package
