+++ src/main/java/fpga/pipeline/Extract.java
@@ -1,4 +1,4 @@
-package fpga;
+package fpga.pipeline;
 
 import java.util.ArrayList;
 import java.util.Collections;
