+++ src/main/java/fpga/pipeline/PipeStream.java
@@ -1,4 +1,4 @@
-package fpga;
+package fpga.pipeline;
 
 import java.util.concurrent.LinkedBlockingQueue;
