+++ src/main/java/fpga/pipeline/Threshold.java
@@ -1,4 +1,4 @@
-package fpga;
+package fpga.pipeline;
 
 import java.awt.image.BufferedImage;
