+++ src/main/java/fpga/pipelinetest/Multiply2Node.java
@@ -1,9 +1,8 @@
 package fpga.pipelinetest;
 
-import fpga.PipeNode;
-import fpga.PipeStream.In;
-import fpga.PipeStream.Out;
-import java.util.List;
+import fpga.pipeline.PipeNode;
+import fpga.pipeline.PipeStream.In;
+import fpga.pipeline.PipeStream.Out;
 
 public class Multiply2Node extends PipeNode <Double, Double>
 {
